test: KIC-in-konnect with gwconf
